+++ python/GafferTest/__init__.py
@@ -162,6 +162,7 @@ from ModuleTest import ModuleTest
 from NumericBookmarkSetTest import NumericBookmarkSetTest
 from NameSwitchTest import NameSwitchTest
 from SpreadsheetTest import SpreadsheetTest
+from ShufflePlugTest import ShufflePlugTest
 
 from IECorePreviewTest import *